+++ dashboard/src/features/orgs/projects/services/components/ServiceForm/components/ImageField/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageField } from './ImageField';
